Assignment3-4 : ColorButton 컴포넌트 분리
